Added method space_sylinder, which replaces the job of points_to_image.
It sorts the sylinder into radius bins sets the density in each bin to
the average density of the stars within that bin. This makes a 1d density
profile for each sylinder.

Also "moved" the sylinder from negative x-values to positive x-values.
It is 100% equivalent, but is easier to read and I do less mathematical
mistakes this way.